Moved options.js (data container for options) into settings, and updated its imports
